Fixed failing UTs
