+++ packages/yourloops/test/unit/components/patient-data/patient-data.hook.test.ts
@@ -38,7 +38,7 @@ import type { ChartPrefs } from '../../../../components/dashboard-widgets/models
 import { createPatient } from '../../common/utils'
 
 jest.mock('../../../../lib/auth')
-jest.mock('../../../../lib/patient/patient.provider')
+jest.mock('../../../../lib/patient/patients.provider')
 describe('usePatientData hook', () => {
   const patientId = 'fakeId'
   const FOURTEEN_DAYS_IN_MS = TimeService.MS_IN_DAY * 14
